Make SwitchHostError and BodyAPIKeyError public
