+++ Sources/AlgoliaSearchClient/Transport/HTTP/HTTPRequest.swift
@@ -79,7 +79,7 @@ class HTTPRequest<ResponseType: Decodable, Output>: AsyncOperation, ResultContai
         throw HTTPTransport.Error.noReachableHosts(intermediateErrors: intermediateErrors)
       }
 
-      let effectiveRequest = try request.setting(host, timeout: timeout)
+      let effectiveRequest = try request.switchingHost(by: host, withBaseTimeout: timeout)
       Logger.loggingService.log(level: .debug, message: description(for: effectiveRequest))
 
       underlyingTask = requester.perform(request: effectiveRequest) { [weak self] (result: IntermediateResult) in
