+++ Sources/AlgoliaSearchClient/Transport/URLSession/URLRequest+Convenience.swift
@@ -118,64 +118,4 @@ extension URLRequest {
 
   }
 
-  static let maxHeaderAPIKeyLength = 500
-
-  private mutating func setAPIKey(_ apiKey: APIKey?) throws {
-
-    guard let apiKeyValue = apiKey?.rawValue, apiKeyValue.count > URLRequest.maxHeaderAPIKeyLength else {
-      self[header: .apiKey] = apiKey?.rawValue
-      return
-    }
-
-    Logger.warning("The API length is > \(URLRequest.maxHeaderAPIKeyLength). Attempt to insert it into request body.")
-
-    guard let body = httpBody else {
-      throw APIKeyError.missingHTTPBody
-    }
-
-    let decodingResult = Result { try JSONDecoder().decode(JSON.self, from: body) }
-    guard case .success(let bodyJSON) = decodingResult else {
-      if case .failure(let error) = decodingResult {
-        throw APIKeyError.bodyDecodingJSONError(error)
-      }
-      return
-    }
-
-    guard case .dictionary(var jsonDictionary) = bodyJSON else {
-      throw APIKeyError.bodyNonKeyValueJSON
-    }
-
-    jsonDictionary["apiKey"] = .string(apiKeyValue)
-
-    let encodingResult = Result { try JSONEncoder().encode(jsonDictionary) }
-    guard case .success(let updatedBody) = encodingResult else {
-      if case .failure(let error) = encodingResult {
-        throw APIKeyError.bodyEncodingJSONError(error)
-      }
-      return
-    }
-
-    httpBody = updatedBody
-  }
-
-  enum APIKeyError: Error, CustomStringConvertible {
-    case missingHTTPBody
-    case bodyDecodingJSONError(Error)
-    case bodyNonKeyValueJSON
-    case bodyEncodingJSONError(Error)
-
-    var description: String {
-      switch self {
-      case .missingHTTPBody:
-        return "Request doesn't contain HTTP body"
-      case .bodyDecodingJSONError(let error):
-        return "HTTP body JSON decoding error: \(error)"
-      case .bodyEncodingJSONError(let error):
-        return "HTTP body JSON encoding error: \(error)"
-      case .bodyNonKeyValueJSON:
-        return "HTTP body JSON is not key-value type"
-      }
-    }
-  }
-
 }
