+++ Tests/AlgoliaSearchClientTests/Unit/Transport/URLRequestConstructionTests.swift
@@ -63,5 +63,19 @@ class URLRequestConstructionTests: XCTestCase {
     XCTAssertNil(request.allHTTPHeaderFields?[HTTPHeaderKey.apiKey.rawValue])
     XCTAssertEqual(json["apiKey"], .string(rawAPIKey))
   }
+  
+  func testSwitchHost() throws {
+    let timeout: TimeInterval = 10
+    let request = URLRequest(command: Command.Custom(method: .post, callType: .write, path: URL(string: "/my/test/path")!, body: nil, requestOptions: nil))
+
+    for index in 0...20 {
+      var host = RetryableHost(url: URL(string: "test\(index).algolia.com")!)
+      host.retryCount = index
+      let requestWithHost = try request.switchingHost(by: host, withBaseTimeout: timeout)
+      XCTAssertEqual(requestWithHost.timeoutInterval, timeout * TimeInterval(index + 1))
+      XCTAssertEqual(requestWithHost.url?.absoluteString, "https://test\(index).algolia.com/my/test/path")
+    }
+
+  }
 
 }
